+++ components/Collections/examples/CollectionsEmptySections.m
@@ -50,7 +50,7 @@ static NSString *const kReusableIdentifierItem = @"itemCellIdentifier";
   for (NSInteger i = 0; i < kSectionCount; i++) {
     NSMutableArray *items = [NSMutableArray array];
     for (NSInteger j = 0; j < kSectionItemCount; j++) {
-      NSString *itemString = [NSString stringWithFormat:@"Section-%zd Item-%zd", i, j];
+      NSString *itemString = [NSString stringWithFormat:@"Section-%ld Item-%ld", (long)i, (long)j];
       [items addObject:itemString];
     }
     [_content addObject:items];
